Wall run bugfixes (#34)
**JIRA Ticket**
https://perigongames.atlassian.net/browse/BFB-50
https://perigongames.atlassian.net/browse/BFB-53
https://perigongames.atlassian.net/browse/BFB-54
https://perigongames.atlassian.net/browse/BFB-56
https://perigongames.atlassian.net/browse/BFB-57
https://perigongames.atlassian.net/browse/BFB-59

**Description**
Bugfixes for wall running, and add in adjustable tuning for it.
For BFB-50 : Check movement is in the positive direction and otherwise stop wall running
For BFB-54 : Added a callback after wall running ends, and a method in the fps player that resets the jump count. 
For BFB-53 : Caches the last collider, and doesn't allow repeated wall running on the same wall. 
For BFB-56 : Set parkourwall layer to walkable, and increase min height a bit
BFB-57 is self explanatory, just added a multiplier on the velocity, and the prefab has a default value of .75 for that
For BFB-59 : End wall run if velocity dips too low