Gursaanj/BFB-195 : SceneSwticher EditorWindow (#108)
**JIRA Ticket**
https://perigongames.atlassian.net/browse/BFB-191

**Description**
Made a SceneSwitcher Editor Window that will allow users to more quickly go between scenes or load them additively on top of current scenes. Only scenes under the `_BForBoss` subfolder are added into this editor window.

**Screenshots / GIFs**
![SceneSwitcher](https://user-images.githubusercontent.com/23490604/152713037-dacd26ed-c609-4c31-9b21-063e5f529806.png)

All scenes under the same parent folder are grouped together. 
The Scene that is highlighted green indicates the current main scene the user is on at the moment.

![SceneSwitcherContextMenu](https://user-images.githubusercontent.com/23490604/152713376-aed72795-cd8f-4a38-b687-7b445389c6be.png)

Anytime a user creates, deletes, or renames a scene, the Editor Window, if open, should update automatically. If for some reason, changes have not been updated on the Editor Window,  a manual `Refresh Scenes` button can be used in the context menu.

![Favoriting Scenes](https://user-images.githubusercontent.com/23490604/152713739-ef61cab0-e1fd-4416-9719-828c12a99b29.gif)


A feature I added, when I originally thought we wanted to have all scenes in the Assets folder included for the Editor Window (i.e. the 20+ ECM2 scenes and such) I created a **Favoriting** Feature where user can set their favourite scenes at the top of the window so they don't have to backtrack each time they want to change the scenes they will be working on the most.
***Note** The favourited scenes are saved via `EditorPrefs` so each individual can have their own favourite scenes without influencing any other users.

I definitely see this feature being way more useful in the coming months when our scene count is much higher.